Move src/lib/resources.js to .serverless_plugins/faultline-resources
